realm_export: Change tarball_size_bytes to PositiveBigIntegerField.
This commit updates `tarball_size_bytes` field from
`PositiveIntegerField` to `PositiveBigIntegerField` to support
larger values.

+++ zerver/models/realms.py
@@ -1368,5 +1368,5 @@ class RealmExport(models.Model):
     acting_user = models.ForeignKey("UserProfile", null=True, on_delete=models.SET_NULL)
     export_path = models.TextField(default=None, null=True)
     sha256sum_hex = models.CharField(default=None, null=True, max_length=64)
-    tarball_size_bytes = models.PositiveIntegerField(default=None, null=True)
+    tarball_size_bytes = models.PositiveBigIntegerField(default=None, null=True)
     stats = models.JSONField(default=None, null=True)
